Ensure original query is performed as-is
Lucene boolean precedence is such that AND takes precedence over OR. So if you searched for ‘role:A OR role:B’, the final query would be ‘role:A OR role:B AND environment:env’. The odd part about this, is that I would expect the return to be all nodes in role:A + nodes in (role:B AND environment:env). Instead, only those nodes in (role:B AND environment:env) are returned.

I honestly cannot explain why this would be the case, but it is. This simple fix will prevent that and produce the expected results.